requirements.txt
fastapi
uvicorn
python-multipart
pillow
pandas
openpyxl
pytesseract
email-validator
aiofiles
jinja2
smtplib
python-dotenv
